backups/2025-06-03/requirements 2.txt
Flask==3.0.2
Flask-Cors==4.0.0
pytest==8.0.2
pytest-asyncio==0.23.5
loguru==0.7.2
python-dotenv==1.0.1
supabase==2.3.4
httpx>=0.24.0,<0.26.0
aiohttp==3.9.3
pydantic==2.5.2
async-timeout==4.0.3
python-dateutil==2.8.2
PyJWT==2.8.0
asgiref==3.7.2 